Refactor some functions and move duplicate code to abstract classes. (#487)
