+++ core/src/main/java/com/alibaba/fescar/core/rpc/netty/TmRpcClient.java
@@ -162,45 +162,12 @@ public final class TmRpcClient extends AbstractRpcRemotingClient {
             @Override
             public void run() {
                 try {
-                    reconnect();
+                    reconnect(transactionServiceGroup);
                 } catch (Exception ignore) {
                     LOGGER.error(ignore.getMessage());
                 }
             }
         }, healthCheckDelay, healthCheckPeriod, TimeUnit.SECONDS);
-    }
-
-    private void reconnect() {
-        List<String> availList = null;
-        try {
-            availList = getAvailServerList(transactionServiceGroup);
-        } catch (Exception exx) {
-            LOGGER.error(exx.getMessage());
-        }
-        if (CollectionUtils.isEmpty(availList)) {
-            LOGGER.error("no available server to connect.");
-            return;
-        }
-        for (String serverAddress : availList) {
-            try {
-                connect(serverAddress);
-            } catch (Exception e) {
-                LOGGER.error(FrameworkErrorCode.NetConnect.errCode,
-                    "can not connect to " + serverAddress + " cause:" + e.getMessage(), e);
-            }
-        }
-    }
-
-    @Override
-    public void channelInactive(ChannelHandlerContext ctx) throws Exception {
-        if (messageExecutor.isShutdown()) {
-            return;
-        }
-        if (LOGGER.isInfoEnabled()) {
-            LOGGER.info("channel inactive:" + ctx.channel());
-        }
-        releaseChannel(ctx.channel(), NetUtil.toStringAddress(ctx.channel().remoteAddress()));
-        super.channelInactive(ctx);
     }
 
     @Override
@@ -210,7 +177,7 @@ public final class TmRpcClient extends AbstractRpcRemotingClient {
         Channel acquireChannel = connect(validAddress);
         Object result = super.sendAsyncRequestWithResponse(validAddress, acquireChannel, msg, timeout);
         if (result instanceof GlobalBeginResponse
-            && ((GlobalBeginResponse)result).getResultCode() == ResultCode.Failed) {
+                && ((GlobalBeginResponse) result).getResultCode() == ResultCode.Failed) {
             LOGGER.error("begin response error,release channel:" + acquireChannel);
             releaseChannel(acquireChannel, validAddress);
         }
@@ -273,14 +240,11 @@ public final class TmRpcClient extends AbstractRpcRemotingClient {
 
     }
 
-    /**
-     * Release channel.
-     *
-     * @param channel       the channel
-     * @param serverAddress the server address
-     */
-    public void releaseChannel(Channel channel, String serverAddress) {
-        if (null == channel || null == serverAddress) { return; }
+    @Override
+    protected void releaseChannel(Channel channel, String serverAddress) {
+        if (null == channel || null == serverAddress) {
+            return;
+        }
         try {
             Object connectLock = channelLocks.get(serverAddress);
             synchronized (connectLock) {
@@ -303,12 +267,6 @@ public final class TmRpcClient extends AbstractRpcRemotingClient {
         }
     }
 
-    /**
-     * Connect channel.
-     *
-     * @param serverAddress the server address
-     * @return the channel
-     */
     @Override
     protected Channel connect(String serverAddress) {
         Channel channelToServer = channels.get(serverAddress);
